+++ lib/dry/schema/primitive_inferrer.rb
@@ -4,7 +4,13 @@ module Dry
   module Schema
     # @api private
     class PrimitiveInferrer < ::Dry::Types::PrimitiveInferrer
-      Compiler = ::Class.new(superclass::Compiler)
+      Compiler = ::Class.new(superclass::Compiler) do
+        # @api private
+        def visit_intersection(node)
+          left, right = node
+          [visit(left), visit(right)].flatten(1)
+        end
+      end
 
       def initialize
         super