+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Break Versioning](https://www.taoensso.com/break-ve
 
 ## [Unreleased]
 
+### Fixed
+
+- Support for intersection types (created with `&` operator) in schema definitions (fixes #494) (@baweaver)
 
 ## [1.14.1] - 2025-03-03